+++ font/dict/type0.go
@@ -171,7 +171,7 @@ func extractCIDFontType0(x *pdf.Extractor, obj pdf.Object) (*CIDFontType0, error
 		if pdf.IsReadError(err) {
 			return nil, err
 		}
-		d.DefaultWidth = float64(dw)
+		d.DefaultWidth = dw
 	} else {
 		d.DefaultWidth = DefaultWidthDefault
 	}
